nginx-0.1.3-RELEASE import
    *) Feature: the ngx_http_autoindex_module and the autoindex directive.

    *) Feature: the proxy_set_x_url directive.

    *) Bugfix: proxy module may get caught in an endless loop when sendfile
       is not used.

+++ src/http/ngx_http_core_module.c
@@ -1012,7 +1012,7 @@ static int ngx_cmp_locations(const void *one, const void *two)
         return 1;
     }
 
-    return rc;
+    return (int) rc;
 }
